requirements.txt
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
autopep8==2.0.4
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.4
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.7
Django==4.2.7
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.4
finnhub-python==2.4.19
fonttools==4.42.1
fredapi==0.5.1
frozendict==2.3.8
html5lib==1.1
idna==3.4
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
MLB-StatsAPI==1.7
multitasking==0.0.11
nest-asyncio==1.5.8
newsapi==0.1.1
newsapi-python==0.2.7
numpy==1.26.0
openpyxl==3.1.2
packaging==23.1
pandas==2.1.1
parso==0.8.3
patsy==0.5.4
peewee==3.16.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
plotly==5.18.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pyarrow==13.0.0
pybaseball==2.2.7
pycodestyle==2.11.0
pycparser==2.21
PyGithub==2.1.1
Pygments==2.16.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post10
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.0
tenacity==8.2.3
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
virtualenv==20.24.5
virtualenvwrapper-win==1.2.7
wcwidth==0.2.6
webencodings==0.5.1
wrapt==1.15.0
yahoofinancials==1.18
yfinance==0.2.30
